owb driver support
Relocate gpio functionality into owb_gpio.h/.c

Add 'struct owb_driver' to store a driver's function pointer.
