+++ owb_gpio.h
@@ -1,0 +1,3 @@
+#pragma once
+
+struct owb_driver* owb_gpio_get_driver();
